[MERGE] merged with develop.

+++ we-re-server/src/reviews/reviews.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { CreateReviewDto } from './dto/create-review.dto';
 import { UpdateReviewDto } from './dto/update-review.dto';
 import { ReviewRepository } from './reviews.repository';
@@ -7,20 +7,24 @@ import {
   ReadReviewAndWebtoonDto,
   ReadReviewDto,
 } from './dto/read-review.dto';
-import { Review } from 'src/entities/review.entity';
 import {
   CustomDataAlreadyExistException,
   CustomDataBaseException,
   CustomNotFoundException,
+  CustomUnauthorziedException,
 } from 'src/utils/custom_exceptions';
 import { LikesService } from 'src/likes/likes.service';
 import { AddAndRemoveLikeDto } from 'src/likes/dto/cud-like.dto';
+import { TagsService } from 'src/tags/tags.service';
+import { TARGET_TYPES } from 'src/utils/types_and_enums';
+import { AddAndRemoveTagRequestDto } from 'src/tags/dto/process-tag.dto';
 
 @Injectable()
 export class ReviewsService {
   constructor(
     private readonly reviewRepository: ReviewRepository,
     private readonly likesService: LikesService,
+    private readonly tagsService: TagsService,
   ) {}
 
   /**
@@ -35,10 +39,14 @@ export class ReviewsService {
         const temp = new ReadReviewAndWebtoonDto(r);
         const addAndRemoveLikeDto = new AddAndRemoveLikeDto(
           userId,
-          'review',
+          TARGET_TYPES.REVIEW,
           temp.id,
         );
         temp.like = await this.likesService.getLikeCount(addAndRemoveLikeDto);
+        temp.tags = await this.tagsService.findTagsByTargetId(
+          TARGET_TYPES.REVIEW,
+          temp.id,
+        );
         return temp;
       }),
     );
@@ -59,6 +67,15 @@ export class ReviewsService {
     const result: ReadReviewAndUserDto[] = queryResult.map(
       (r) => new ReadReviewAndUserDto(r),
     );
+    await Promise.all(
+      result.map(
+        async (r) =>
+          (r.tags = await this.tagsService.findTagsByTargetId(
+            TARGET_TYPES.REVIEW,
+            r.id,
+          )),
+      ),
+    );
     return result;
   }
 
@@ -67,11 +84,14 @@ export class ReviewsService {
    * @param createReviewDto user id and webtoon id
    * @returns {Promise<Review>}
    */
-  async createReview(createReviewDto: CreateReviewDto): Promise<Review> {
+  async createReview(
+    userId: number,
+    createReviewDto: CreateReviewDto,
+  ): Promise<ReadReviewDto> {
     const alreadyReview = await this.reviewRepository.findOne({
       where: {
-        user: { id: +createReviewDto.userId },
-        webtoon: { id: +createReviewDto.webtoonId },
+        user: { id: userId },
+        webtoon: { id: createReviewDto.webtoonId },
       },
     });
     if (alreadyReview) {
@@ -80,14 +100,26 @@ export class ReviewsService {
       );
     }
     const queryResult = await this.reviewRepository.createReview(
+      userId,
       createReviewDto,
     );
     const id = queryResult.identifiers[0].id;
     if (!id) {
       throw new CustomDataBaseException('create is not worked.');
     }
-    const result = await this.reviewRepository.findOneBy(id);
-    return result;
+    const result = await this.reviewRepository.findOneById(id);
+    const readReviewDto = new ReadReviewDto(result);
+    const { tags } = createReviewDto;
+    const addAndRemoveTagRequestDto = new AddAndRemoveTagRequestDto(
+      TARGET_TYPES.REVIEW,
+      id,
+      tags,
+    );
+    readReviewDto.tags = await this.tagsService.addAndRemoveTag(
+      addAndRemoveTagRequestDto,
+    );
+
+    return readReviewDto;
   }
 
   /**
@@ -95,25 +127,45 @@ export class ReviewsService {
    * @param updateReviewDto id and fields
    * @returns {Promise<Review>}
    */
-  async updateReview(updateReviewDto: UpdateReviewDto): Promise<Review> {
+  async updateReview(
+    userId: number,
+    updateReviewDto: UpdateReviewDto,
+  ): Promise<ReadReviewDto> {
+    const { tags, id, ...tempUpdateReviewDto } = updateReviewDto;
+    await this.checkReviewOwner(id, userId);
     const queryResult = await this.reviewRepository.update(
-      updateReviewDto.id,
-      updateReviewDto,
+      id,
+      tempUpdateReviewDto,
     );
     if (!queryResult.affected) {
       throw new CustomNotFoundException('id');
     }
-    const result = await this.reviewRepository.findOneBy({
-      id: updateReviewDto.id,
-    });
-    return result;
+    const result = await this.reviewRepository.findOneById(id);
+    const readReviewDto = new ReadReviewDto(result);
+    const addAndRemoveTagRequestDto = new AddAndRemoveTagRequestDto(
+      TARGET_TYPES.REVIEW,
+      result.id,
+      tags,
+    );
+    readReviewDto.tags = await this.tagsService.addAndRemoveTag(
+      addAndRemoveTagRequestDto,
+    );
+
+    return readReviewDto;
   }
 
-  async deleteReview(id: number): Promise<void> {
+  async deleteReview(id: number, userId: number): Promise<void> {
+    await this.checkReviewOwner(id, userId);
     const queryResult = await this.reviewRepository.delete(id);
     if (!queryResult) {
       throw new CustomNotFoundException('id');
     }
     return;
+  }
+
+  async checkReviewOwner(id: number, userId: number): Promise<void> {
+    const queryResult = await this.reviewRepository.findOneById(id);
+    if (queryResult.user_id == userId)
+      throw new CustomUnauthorziedException("you can't change review");
   }
 }
